LayoutTests/imported/w3c/web-platform-tests/file-system-access/sandboxed_FileSystemSyncAccessHandle-flush.https.tentative.worker-expected.txt


PASS Test flush on an empty file.
PASS SyncAccessHandle.read returns bytes written by SyncAccessHandle.write after SyncAccessHandle.flush
PASS SyncAccessHandle.read fails when there is a pending SyncAccessHandle.flush
PASS SyncAccessHandle.write fails when there is a pending SyncAccessHandle.flush
